requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
asyncpg
pydantic>=2.0
python-dotenv
